backend/pkspotapi/serializers.py
from rest_framework import serializers
from pkspotapp.models import Spot, MyUser

class SpotSerializer(serializers.ModelSerializer):
    class Meta:
        model = Spot
        fields = '__all__'


class MyUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = MyUser
        fields = '__all__'
