backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.12
django-cors-headers==3.11.0
django-mysql==4.5.0
djangorestframework==3.12.4
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
jsonpickle==2.2.0
MarkupSafe==2.1.1
mysqlclient==2.1.0
oauthlib==3.2.0
pandas==1.4.2
Pillow==9.1.0
pycparser==2.21
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
